# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: React Sheets Workbook API
  version: 2026-08-24
servers:
  - url: /
paths:
...
  /api/workbooks/{unitId}:
    get:
      operationId: getWorkbookSummary
      parameters:
        - $ref: '#/components/parameters/UnitId'
      responses:
        '200':
          description: Workbook summary for the authenticated actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkbookSummary'
  /api/workbooks:
    get:
      operationId: listWorkbooks
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Bounded workbook catalog page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkbookPage'
    post:
      operationId: createWorkbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkbookRequest'
      responses:
        '201':
          description: Workbook snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
  /api/workbooks/{unitId}/operations:
    post:
      operationId: commitOperation
      parameters:
        - $ref: '#/components/parameters/UnitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationEnvelope'
      responses:
        '201':
          description: Committed operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitResponse'
  /api/workbooks/{unitId}/operations/{operationId}:
    get:
      operationId: getOperationResult
      parameters:
        - $ref: '#/components/parameters/UnitId'
        - name: operationId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Previously committed operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
  /api/workbooks/{unitId}/revisions:
    get:
      operationId: listRevisions
      parameters:
        - $ref: '#/components/parameters/UnitId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Bounded revision page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionPage'
  /api/workbooks/{unitId}/assets/{assetId}:
    put:
      operationId: putWorkbookAsset
      parameters:
        - $ref: '#/components/parameters/UnitId'
        - name: assetId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: X-Content-SHA256
          in: header
          required: true
          schema: { type: string, pattern: '^[a-f0-9]{64}$' }
        - name: X-Asset-MimeType
          in: header
          required: true
          schema: { type: string, pattern: '^image/' }
        - name: X-Asset-Width
          in: header
          required: false
          schema: { type: integer, minimum: 1 }
        - name: X-Asset-Height
          in: header
          required: false
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/octet-stream: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: Content-addressed image metadata
          content:
            application/json: { schema: { $ref: '#/components/schemas/AssetMetadata' } }
    get:
      operationId: getWorkbookAsset
      parameters:
        - $ref: '#/components/parameters/UnitId'
        - name: assetId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Image bytes
          headers:
            X-Content-SHA256: { schema: { type: string } }
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
    delete:
      operationId: deleteWorkbookAsset
      parameters:
        - $ref: '#/components/parameters/UnitId'
        - name: assetId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '204': { description: Asset released }
  /api/workbooks/{unitId}/assets/reconcile:
    post:
      operationId: reconcileWorkbookAssets
      parameters: [{ $ref: '#/components/parameters/UnitId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetIds]
              properties:
                assetIds: { type: array, items: { type: string }, uniqueItems: true }
      responses:
        '204': { description: Unreferenced assets released }
...
components:
  parameters:
    UnitId:
      name: unitId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
  schemas:
...
    WorkbookSummary:
      type: object
      required: [unitId, name, revision, updatedAt, lifecycle]
      properties:
        unitId: { type: string, minLength: 1, maxLength: 200 }
        name: { type: string }
        revision: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
        role: { enum: [owner, editor, commenter, viewer] }
        ownerSubject: { type: [string, 'null'] }
        spaceId: { type: [string, 'null'] }
        folderId: { type: [string, 'null'] }
        locationPath: { type: array, items: { type: string } }
        spaceName: { type: [string, 'null'] }
        sourceFileName: { type: [string, 'null'] }
        storageLocation: { enum: [local, remote, mirrored] }
        syncStatus: { enum: [synced, syncing, pending, offline, conflict, error] }
        lifecycle: { enum: [active, trashed] }
        source: { enum: [native, document-import] }
        favorite: { type: boolean }
        lastOpenedAt: { type: [string, 'null'], format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
...
    AssetMetadata:
      type: object
      required: [schema, unitId, assetId, contentHash, mimeType, byteLength, updatedAt]
      properties:
        schema: { const: AssetRef }
        unitId: { type: string }
        assetId: { type: string }
        contentHash: { type: string, pattern: '^[a-f0-9]{64}$' }
        mimeType: { type: string, pattern: '^image/' }
        byteLength: { type: integer, minimum: 1 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        updatedAt: { type: string, format: date-time }
    WorkbookSnapshot:
      type: object
      required: [schema, version, unitId, name, sheets, dataModel, dimensionMetrics]
      properties:
        schema: { const: WorkbookSnapshot }
        version: { const: 9 }
        unitId: { type: string, minLength: 1, maxLength: 200 }
        name: { type: string, minLength: 1, maxLength: 255 }
        dimensionMetrics:
          type: object
          required: [normalFontFamily, normalFontSizePx, maximumDigitWidthPx]
          properties:
            normalFontFamily: { type: string, minLength: 1 }
            normalFontSizePx: { type: number, exclusiveMinimum: 0 }
            maximumDigitWidthPx: { type: number, exclusiveMinimum: 0 }
        sheets: { type: array, minItems: 1, items: { $ref: '#/components/schemas/WorksheetSnapshot' } }
        dataModel:
          type: object
          required: [sources, tables, relationships, views]
          properties:
            sources: { type: array, items: { type: object } }
            tables: { type: array, items: { type: object } }
            relationships: { type: array, items: { type: object } }
            views: { type: array, items: { type: object } }
        cellStyleTemplates: { type: array, items: { type: object } }
    WorksheetPane:
      oneOf:
        - type: object
          required: [kind]
          properties: { kind: { const: none } }
          additionalProperties: false
        - type: object
          required: [kind, xSplit, ySplit, startRow, startColumn, state]
          properties:
            kind: { enum: [frozen, split] }
            xSplit: { type: number, minimum: 0 }
            ySplit: { type: number, minimum: 0 }
            startRow: { type: integer, minimum: 0 }
            startColumn: { type: integer, minimum: 0 }
            activePane: { enum: [topLeft, topRight, bottomLeft, bottomRight] }
            state: { enum: [frozen, frozenSplit, split] }
    AutoFilter:
      type: object
      required: [sheetId, range, columns]
      properties:
        sheetId: { type: string, minLength: 1 }
        range: { type: object }
        columns: { type: object, additionalProperties: { type: object } }
        sortState: { type: object }
        additionalProperties: false
    WorksheetSnapshot:
      type: object
      required: [id, name, rowCount, columnCount, cells, merges, pane, pivots, sparklines, drawings, drawingPayloads, review, defaultRowHeightPx, defaultColumnWidthPx]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        rowCount: { type: integer, minimum: 1, maximum: 1048576 }
        columnCount: { type: integer, minimum: 1, maximum: 16384 }
        cells: { type: object, additionalProperties: { type: object } }
        merges: { type: array, items: { type: object } }
        pane: { $ref: '#/components/schemas/WorksheetPane' }
        autoFilter: { $ref: '#/components/schemas/AutoFilter' }
        pivots: { type: array, items: { type: object } }
        sparklines: { type: array, items: { type: object } }
        drawings: { type: array, items: { type: object } }
        drawingPayloads: { type: object, additionalProperties: { type: object } }
        review:
          type: object
          required: [notesByCell, notesById, threadIdsByCell, threadsById]
          properties:
            notesByCell: { type: object, additionalProperties: { type: string, minLength: 1 } }
            notesById: { type: object, additionalProperties: { type: object } }
            threadIdsByCell: { type: object, additionalProperties: { type: array, items: { type: string, minLength: 1 } } }
            threadsById: { type: object, additionalProperties: { type: object } }
          additionalProperties: false
        defaultRowHeightPx: { type: number, exclusiveMinimum: 0 }
        defaultColumnWidthPx: { type: number, exclusiveMinimum: 0 }
        rowHeightsPx: { type: object, additionalProperties: { type: number, exclusiveMinimum: 0 } }
        columnWidthsPx: { type: object, additionalProperties: { type: number, exclusiveMinimum: 0 } }
    CreateWorkbookRequest:
      type: object
      required: [unitId, name, snapshot]
      properties:
        unitId: { type: string, minLength: 1, maxLength: 200 }
        name: { type: string, minLength: 1, maxLength: 255 }
        snapshot: { $ref: '#/components/schemas/WorkbookSnapshot' }
        spaceId: { type: [string, 'null'] }
        folderId: { type: [string, 'null'] }
    SnapshotResponse:
      type: object
      required: [unitId, snapshot, revision, checksum]
      properties:
        unitId: { type: string }
        snapshot: { $ref: '#/components/schemas/WorkbookSnapshot' }
        revision: { type: integer, minimum: 0 }
        checksum: { type: string }
    OperationEnvelope:
      type: object
      required: [schema, operationId, unitId, clientSessionId, clientSequence, baseRevision, mutations, createdAt]
      properties:
        schema: { const: OperationEnvelope }
        operationId: { type: string, minLength: 1 }
        unitId: { type: string, minLength: 1 }
        clientSessionId: { type: string, minLength: 1, maxLength: 200 }
        clientSequence: { type: integer, minimum: 1 }
        baseRevision: { type: integer, minimum: 0 }
        mutations: { type: array, minItems: 1, items: { type: object } }
        createdAt: { type: string, format: date-time }
    CommitResponse:
      type: object
      required: [operation]
      properties:
        operation: { $ref: '#/components/schemas/CommittedOperationEnvelope' }
    OperationResult:
      type: object
      required: [operation, committed]
      properties:
        operation: { $ref: '#/components/schemas/CommittedOperationEnvelope' }
        committed: { type: boolean }
    CommittedOperationEnvelope:
      type: object
      required: [clientSessionId, schema, operationId, unitId, actorId, origin, clientSequence, baseRevision, revision, mutations, createdAt, committedAt]
      properties:
        clientSessionId: { type: string, minLength: 1, maxLength: 200 }
        schema: { const: OperationEnvelope }
        operationId: { type: string, minLength: 1 }
        unitId: { type: string, minLength: 1, maxLength: 200 }
        actorId: { type: string, minLength: 1, maxLength: 200 }
        origin: { enum: [client, system] }
        clientSequence: { type: integer, minimum: 1 }
        baseRevision: { type: integer, minimum: 0 }
        revision: { type: integer, minimum: 1 }
        mutations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CommittedOperationMutation'
        createdAt: { type: string, format: date-time }
        committedAt: { type: string, format: date-time }
        intent: { type: [object, 'null'] }
    CommittedOperationMutation:
      type: object
      required: [id, sheetId, params, affectedRanges]
      properties:
        id: { type: string, minLength: 1 }
        sheetId: { type: string, minLength: 1 }
        params: { type: object }
        affectedRanges:
          type: array
          items:
            $ref: '#/components/schemas/RangeRef'
    RangeRef:
      type: object
      required: [sheetId, startRow, endRow, startColumn, endColumn]
      properties:
        sheetId: { type: string, minLength: 1 }
        startRow: { type: integer, minimum: 0 }
        endRow: { type: integer, minimum: 0 }
        startColumn: { type: integer, minimum: 0 }
        endColumn: { type: integer, minimum: 0 }
    WorkbookPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { type: object } }
        nextCursor: { type: [string, 'null'] }
    RevisionPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { type: object } }
        nextCursor: { type: [string, 'null'] }
...